+++ next.config.js
@@ -16,3 +16,4 @@ module.exports = withPWA({
   },
   nextConfig
 })
+
